refactor(ioxide): Infrastructure/Endpoints, matching the Internal engine
Hosting/ held the server, its host and the endpoint types in one flat folder,
under a name no other engine uses. It now mirrors Engine/Internal:

    Infrastructure/
      IoxideServer.cs  .Tcp.cs  .Tcp.Tls.cs  .Quic.cs
      IoxideServerHost.cs
      Endpoints/
        EndPoint.cs
        EndPointCollection.cs

IoxideEndPoint and IoxideEndPoints lose the prefix the namespace already carries,
and become EndPoint and EndPointCollection - the names the Internal engine gives
the same two things. Both are internal now, as they are there: neither was ever
reachable except through IEndPoint and IEndPointCollection.

Namespaces follow the folders, so the server files move to
GenHTTP.Engine.Ioxide.Infrastructure and the endpoint types sit one below.
