+++ Engine/Ioxide/GenHTTP.Engine.Ioxide.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     
-    <TargetFrameworks>net11.0</TargetFrameworks>
+    <TargetFrameworks>net11.0;net10.0</TargetFrameworks>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     
   </PropertyGroup>
